+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Diskover Web Change Log
 
+# [1.5.0-rc8] - 2018-06-09
+### added
+- better warning messages that workers are still calculating directory sizes to analytics pages and dashboard
+### fixed
+- top 10 directories on dashboard to match same top 10 in top 50 analytics page
+
 # [1.5.0-rc7] - 2018-05-30
 ### added
 - heatmap, dupes, hardlinks to analytics button dropdown list on search results and file/dir view pages